Refactor and standardization of the build process.
