+++ bindings/CMakeLists.txt
@@ -4,5 +4,18 @@ if(NOT pybind11_FOUND)#
   return()
 endif()
 
+## Python configuration file (build tree), the lib in the source dir (easier while developping .py files)
+file(WRITE "${CMAKE_BINARY_DIR}/etc/sofa/python.d/plugin.SofaPython3.bindings" "${CMAKE_BINARY_DIR}/${SP3_PYTHON_PACKAGES_DIRECTORY}")
+
+## Python configuration file (install tree)
+file(WRITE "${CMAKE_CURRENT_BINARY_DIR}/installed-bindings-config"
+    ${CMAKE_INSTALL_PREFIX}/${SP3_PYTHON_PACKAGES_DIRECTORY})
+
+
+install(FILES "${CMAKE_CURRENT_BINARY_DIR}/installed-bindings-config"
+    DESTINATION "etc/sofa/python.d"
+    RENAME "plugin.SofaPython3.bindings")
+
 add_subdirectory(Sofa)
 add_subdirectory(SofaRuntime)
+add_subdirectory(SofaTypes)
