--- bindings/Sofa/package/__Sofa_Types__/__init__.py
@@ -1,6 +1,0 @@
-"""
-
-"""
-from __Sofa_Types__.Vec3 import *
-from __Sofa_Types__.RGBAColor import *
-
