+++ bindings/SofaTypes/package/__init__.py
@@ -1,0 +1,7 @@
+"""
+SofaType package
+-----------------------
+
+"""
+
+from .@MODULE_NAME@  import *
