+++ bindings/SofaTypes/src/SofaPython3/SofaTypes/Binding_BoundingBox.h
