+++ bindings/SofaTypes/src/SofaPython3/SofaTypes/Binding_Quat.cpp
